# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.CommitCertRequest:
    description: 生成存证证书的请求体
    properties:
      evidence_level:
        description: 存证级别：internal/compliance/legal（或旧格式 L1/L2/L3）
        example: internal
        type: string
      trace_id:
        description: 追踪 ID
        example: trc_abc123
        type: string
    required:
    - trace_id
    type: object
...
  api.GenerateProofRequest:
    description: 生成最小披露证明的请求体
    properties:
      disclose_events:
        description: 要披露的事件索引（如 [0, 2]）
        items:
          type: integer
        type: array
      disclose_fields:
        description: 要披露的字段（如 ["output_hash"]）
        items:
          type: string
        type: array
    type: object
...
  api.IngestEventResult:
    description: 单个事件的写入结果
    properties:
      error:
        description: 错误信息（如有）
        example: duplicate event_id
        type: string
      event_hash:
        description: 事件哈希
        example: sha256:abcd1234...
        type: string
      event_id:
        description: 事件 ID
        example: evt_abc123
        type: string
    type: object
  api.IngestEventsRequest:
    type: object
  api.IngestEventsResponse:
    description: 事件写入的响应结果
    properties:
      results:
        description: 各事件的写入结果
        items:
          $ref: '#/definitions/api.IngestEventResult'
        type: array
      success:
        description: 是否成功
        example: true
        type: boolean
    type: object
...
  api.VerifyCertRequest:
    description: 验证存证证书的请求体
    properties:
      cert_id:
        description: 证书 ID
        example: cert_abc123
        type: string
      root_hash:
        description: Merkle 根哈希
        example: sha256:abcd1234
        type: string
    type: object
  cert.AnchorProof:
    properties:
      anchor_id:
        type: string
      anchor_timestamp:
        type: string
      anchor_type:
        description: local, worm, blockchain
        type: string
      blockchain:
        $ref: '#/definitions/cert.BlockchainProof'
      object_key:
        type: string
      storage_provider:
        type: string
    type: object
  cert.BlockchainProof:
    properties:
      block_height:
        type: integer
      chain_id:
        type: string
      contract_address:
        type: string
      tx_hash:
        type: string
    type: object
  cert.CertMetadata:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      evidence_level:
        $ref: '#/definitions/cert.EvidenceLevel'
      public_key:
        description: Ed25519公钥（用于验证签名）
        type: string
      tenant_id:
        type: string
    type: object
  cert.Certificate:
    properties:
      anchor_proof:
        allOf:
        - $ref: '#/definitions/cert.AnchorProof'
        description: 锚定证明
      cert_id:
        description: 存证标识
        type: string
      cert_version:
        type: string
      disclosure_policy:
        allOf:
        - $ref: '#/definitions/cert.DisclosurePolicy'
        description: 披露策略
      event_hashes:
        description: 事件哈希列表
        items:
          type: string
        type: array
      merkle_tree:
        allOf:
        - $ref: '#/definitions/merkle.Tree'
        description: Merkle树
      metadata:
        allOf:
        - $ref: '#/definitions/cert.CertMetadata'
        description: 元信息
      root_hash:
        type: string
      schema_version:
        type: string
      time_proof:
        allOf:
        - $ref: '#/definitions/cert.TimeProof'
        description: 时间证明
      trace_id:
        description: 关联的trace
        type: string
    type: object
  cert.DisclosedEvent:
    properties:
      disclosed_fields:
        additionalProperties: true
        type: object
      event_hash:
        type: string
      event_index:
        type: integer
      event_type:
        type: string
    type: object
  cert.DisclosurePolicy:
    properties:
      allowed_fields:
        items:
          type: string
        type: array
      encrypted_fields:
        items:
          type: string
        type: array
      masked_fields:
        items:
          type: string
        type: array
      policy_id:
        type: string
    type: object
  cert.EventMerkleProof:
    properties:
      event_hash:
        type: string
      event_index:
        type: integer
      proof_path:
        items:
          $ref: '#/definitions/merkle.ProofNode'
        type: array
    type: object
  cert.EvidenceLevel:
    enum:
    - internal
    - compliance
    - legal
    - internal
    - compliance
    - legal
    type: string
    x-enum-comments:
      EvidenceLevelL1: Legacy L1 -> internal
      EvidenceLevelL2: Legacy L2 -> compliance
      EvidenceLevelL3: Legacy L3 -> legal
    x-enum-descriptions:
    - Legacy L1 -> internal
    - Legacy L2 -> compliance
    - Legacy L3 -> legal
    - ""
    - ""
    - ""
    x-enum-varnames:
    - EvidenceLevelL1
    - EvidenceLevelL2
    - EvidenceLevelL3
    - EvidenceLevelInternal
    - EvidenceLevelCompliance
    - EvidenceLevelLegal
  cert.MinimalDisclosureProof:
    properties:
      anchor_proof:
        allOf:
        - $ref: '#/definitions/cert.AnchorProof'
        description: 锚定证明
      cert_id:
        description: 证明目标
        type: string
      disclosed_events:
        description: 披露的事件
        items:
          $ref: '#/definitions/cert.DisclosedEvent'
        type: array
      merkle_proofs:
        description: Merkle证明
        items:
          $ref: '#/definitions/cert.EventMerkleProof'
        type: array
      root_hash:
        type: string
      schema_version:
        type: string
      time_proof:
        allOf:
        - $ref: '#/definitions/cert.TimeProof'
        description: 时间证明
      verification_instructions:
        allOf:
        - $ref: '#/definitions/cert.VerificationInstructions'
        description: 验证指引
    type: object
  cert.TimeProof:
    properties:
      proof_type:
        description: local, ntp, tsa
        type: string
      signature:
        type: string
      timestamp:
        type: string
      tsa_authority:
        type: string
      tsa_cert_chain:
        items:
          type: string
        type: array
      tsa_token:
        type: string
    type: object
  cert.VerificationInstructions:
    properties:
      verifier_url:
        type: string
      verify_command:
        type: string
    type: object
...
  gateway.ChatCompletionRequest:
    properties:
      max_tokens:
        type: integer
      messages:
        items:
          $ref: '#/definitions/gateway.ChatMessage'
        type: array
      model:
        type: string
      seed:
        type: integer
      stream:
        type: boolean
      temperature:
        type: number
      top_p:
        type: number
    type: object
  gateway.ChatCompletionResponse:
    properties:
      choices:
        items:
          $ref: '#/definitions/gateway.Choice'
        type: array
      created:
        type: integer
      id:
        type: string
      model:
        type: string
      object:
        type: string
      usage:
        $ref: '#/definitions/gateway.Usage'
    type: object
  gateway.ChatMessage:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
  gateway.Choice:
    properties:
      finish_reason:
        type: string
      index:
        type: integer
      message:
        $ref: '#/definitions/gateway.ChatMessage'
    type: object
  gateway.Usage:
    properties:
      completion_tokens:
        type: integer
      prompt_tokens:
        type: integer
      total_tokens:
        type: integer
    type: object
  merkle.ProofNode:
    properties:
      hash:
        type: string
      position:
        description: '"left" or "right"'
        type: string
    type: object
  merkle.Tree:
    properties:
      algorithm:
        type: string
      leaves:
        items:
          type: string
        type: array
      nodes:
        items:
          items:
            type: string
          type: array
        type: array
      root:
        type: string
    type: object
...
host: localhost:8006
info:
  contact:
    email: support@aitrace.cc
    name: AI-Trace Support
    url: https://aitrace.cc/support
  description: |-
    AI-Trace - Enterprise AI Decision Audit & Compliance Platform

    AI-Trace 提供零代码侵入的 AI 调用追踪、存证和审计能力。
    支持 OpenAI、Claude、Ollama 等多种模型，用户可透传自己的 API Key，服务不存储任何敏感信息。

    ## 特性
    - **零侵入集成**: 只需替换 API endpoint，无需修改代码
    - **API Key 透传**: 用户的 API Key 直接透传到上游，不存储
    - **自定义代理**: 支持用户指定自己的代理服务器
    - **多级存证**: internal(内部审计) / compliance(合规) / legal(法律效力)
    - **最小披露**: 基于 Merkle 树的选择性披露证明
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://aitrace.cc/terms
  title: AI-Trace API
  version: 0.2.0
paths:
  /certs/{cert_id}:
    get:
      description: 根据证书 ID 获取存证证书的完整信息
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 证书 ID
        in: path
        name: cert_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 证书详情
          schema:
            $ref: '#/definitions/cert.Certificate'
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 证书不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取存证证书详情
      tags:
      - Certificates
  /certs/{cert_id}/prove:
    post:
      consumes:
      - application/json
      description: |-
        生成只包含必要信息的证明，支持选择性披露

        ## 使用场景
        - 向第三方证明 AI 决策过程，而不暴露完整对话内容
        - 选择性披露特定事件和字段
        - 生成可独立验证的 Merkle 证明
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 证书 ID
        in: path
        name: cert_id
        required: true
        type: string
      - description: 证明请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.GenerateProofRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 最小披露证明
          schema:
            $ref: '#/definitions/cert.MinimalDisclosureProof'
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 证书不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 生成最小披露证明
      tags:
      - Certificates
  /certs/commit:
    post:
      consumes:
      - application/json
      description: |-
        为指定的追踪生成存证证书，支持多级存证

        ## 存证级别
        - **internal** (L1): Ed25519 签名，适合内部审计
        - **compliance** (L2): WORM 存储 + TSA 时间戳，适合合规要求
        - **legal** (L3): 区块链锚定，适合法律效力要求
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 存证请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CommitCertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 存证证书信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 追踪不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 生成存证证书
      tags:
      - Certificates
  /certs/search:
    get:
      description: 获取当前租户的存证证书列表
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量（最大100）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 证书列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 搜索存证证书
      tags:
      - Certificates
  /certs/verify:
    post:
      consumes:
      - application/json
      description: |-
        验证存证证书的完整性和有效性

        ## 验证项目
        - **hash_integrity**: 事件哈希与 Merkle 根匹配
        - **signature**: Ed25519 数字签名验证
        - **time_proof**: 时间证明验证
        - **anchor_proof**: 锚定证明验证（WORM/区块链）
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 验证请求（cert_id 或 root_hash 二选一）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.VerifyCertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 验证结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 证书不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 验证存证证书
      tags:
      - Certificates
  /chat/completions:
    post:
      consumes:
      - application/json
      description: |-
        代理 OpenAI/Claude/Ollama 聊天完成请求，同时记录追踪事件

        ## API Key 透传
        用户的 API Key 通过以下方式传递，**不会被存储**：
        - `X-Upstream-API-Key` header（推荐）
        - `Authorization: Bearer sk-...` header（兼容模式）

        ## 自定义代理
        可通过 `X-Upstream-Base-URL` 指定自己的代理服务器，避免 IP 封禁风险
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 上游 API Key（OpenAI/Claude，透传不存储）
        in: header
        name: X-Upstream-API-Key
        type: string
      - description: 自定义上游代理 URL
        in: header
        name: X-Upstream-Base-URL
        type: string
      - description: 追踪 ID（可选，自动生成）
        in: header
        name: X-Trace-ID
        type: string
      - description: 会话 ID
        in: header
        name: X-Session-ID
        type: string
      - description: 业务 ID
        in: header
        name: X-Business-ID
        type: string
      - description: 聊天请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gateway.ChatCompletionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 聊天响应
          headers:
            X-Latency-Ms:
              description: 延迟毫秒数
              type: string
            X-Trace-ID:
              description: 追踪 ID
              type: string
          schema:
            $ref: '#/definitions/gateway.ChatCompletionResponse'
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 聊天完成（OpenAI 兼容）
      tags:
      - Gateway
...
  /events/{event_id}:
    get:
      description: 根据事件 ID 获取单个事件的详细信息
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 事件 ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 事件详情
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 事件不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取事件详情
      tags:
      - Events
  /events/ingest:
    post:
      consumes:
      - application/json
      description: 将事件批量写入追踪系统，用于 SDK 直接上报场景
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 租户 ID
        in: header
        name: X-Tenant-ID
        type: string
      - description: 事件列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.IngestEventsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 写入结果
          schema:
            $ref: '#/definitions/api.IngestEventsResponse'
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 批量写入事件
      tags:
      - Events
  /events/search:
    get:
      description: 根据条件搜索追踪事件
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: 追踪 ID
        in: query
        name: trace_id
        type: string
      - description: 事件类型（INPUT/MODEL/OUTPUT）
        in: query
        name: event_type
        type: string
      - description: 开始时间（RFC3339 格式）
        in: query
        name: start_time
        type: string
      - description: 结束时间（RFC3339 格式）
        in: query
        name: end_time
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 事件列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求无效
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 搜索事件
      tags:
      - Events
...
  /getting-started:
    get:
      description: 返回 AI-Trace 快速入门指南，帮助新用户了解如何使用平台
      parameters:
      - description: AI-Trace 平台 API Key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 快速入门指南
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 快速入门
      tags:
      - System
...
securityDefinitions:
  ApiKeyAuth:
    description: AI-Trace 平台 API Key（用于认证）
    in: header
    name: X-API-Key
    type: apiKey
  UpstreamApiKey:
    description: 用户的上游 API Key（OpenAI/Claude，透传不存储）
    in: header
    name: X-Upstream-API-Key
    type: apiKey
swagger: "2.0"